--- include/models/FiguresOfMerit.h
@@ -1,38 +1,0 @@
-//////////////////////////////////////////////////////////////////////////////////////
-//       _______.  ______  __       ___      .__   __. .___________. __  ___   ___  //
-//      /       | /      ||  |     /   \     |  \ |  | |           ||  | \  \ /  /  //
-//     |   (----`|  ,----'|  |    /  ^  \    |   \|  | `---|  |----`|  |  \  V  /   //
-//      \   \    |  |     |  |   /  /_\  \   |  . `  |     |  |     |  |   >   <    //
-//  .----)   |   |  `----.|  |  /  _____  \  |  |\   |     |  |     |  |  /  .  \   //
-//  |_______/     \______||__| /__/     \__\ |__| \__|     |__|     |__| /__/ \__\  //
-//                                                                                  //
-//  Originally developed by D. Pizzocri & T. Barani                                 //
-//                                                                                  //
-//  Version: 2.0                                                                    //
-//  Year: 2022                                                                      //
-//  Authors: D. Pizzocri, G. Zullo.                                                 //
-//                                                                                  //
-//////////////////////////////////////////////////////////////////////////////////////
-
-#ifndef FIGURES_OF_MERIT_H
-#define FIGURES_OF_MERIT_H
-
-#include "SciantixVariableDeclaration.h"
-#include "MapSciantixVariable.h"
-#include "PhysicsVariableDeclaration.h"
-#include "MapPhysicsVariable.h"
-#include "ConstantNumbers.h"
-#include "HistoryVariableDeclaration.h"
-#include "MapHistoryVariable.h"
-#include "MatrixDeclaration.h"
-#include "MapMatrix.h"
-#include <cmath>
-
-void FiguresOfMerit();
-/**
- * @brief This routines computes the value of sciantix_variable associated to figures of merit
- * that are of interest in the output.txt file.
- * 
- */
-
-#endif // FIGURES_OF_MERIT_H
