--- include/models/GrainBoundarySweeping.h
@@ -1,21 +1,0 @@
-//////////////////////////////////////////////////////////////////////////////////////
-//       _______.  ______  __       ___      .__   __. .___________. __  ___   ___  //
-//      /       | /      ||  |     /   \     |  \ |  | |           ||  | \  \ /  /  //
-//     |   (----`|  ,----'|  |    /  ^  \    |   \|  | `---|  |----`|  |  \  V  /   //
-//      \   \    |  |     |  |   /  /_\  \   |  . `  |     |  |     |  |   >   <    //
-//  .----)   |   |  `----.|  |  /  _____  \  |  |\   |     |  |     |  |  /  .  \   //
-//  |_______/     \______||__| /__/     \__\ |__| \__|     |__|     |__| /__/ \__\  //
-//                                                                                  //
-//  Originally developed by D. Pizzocri & T. Barani                                 //
-//                                                                                  //
-//  Version: 2.0                                                                    //
-//  Year: 2022                                                                      //
-//  Authors: D. Pizzocri, G. Zullo.                                                 //
-//                                                                                  //
-//////////////////////////////////////////////////////////////////////////////////////
-
-#include "SciantixVariableDeclaration.h"
-#include "ModelDeclaration.h"
-#include "MapSciantixVariable.h"
-
-void GrainBoundarySweeping();
