+++ src/models/GrainBoundarySweeping.C
@@ -14,7 +14,7 @@
 //                                                                                  //
 //////////////////////////////////////////////////////////////////////////////////////
 
-#include "GrainBoundarySweeping.h"
+#include "Simulation.h"
 
 /**
  * @brief This routine defines the model the grain-boundary sweeping.
@@ -22,44 +22,101 @@
  * 
  */
 
-void GrainBoundarySweeping()
+void Simulation::GrainBoundarySweeping()
 {
-	model.emplace_back();
-	int model_index = int(model.size()) - 1;
+    model.emplace_back();
+    int model_index = int(model.size()) - 1;
 
-	model[model_index].setName("Grain-boundary sweeping");
+    model[model_index].setName("Grain-boundary sweeping");
 
-	switch (int(input_variable[iv["iGrainBoundarySweeping"]].getValue()))
-	{
-	case 0:
-	{
-		std::vector<double> parameter;
-		parameter.push_back(0.0);
-		model[model_index].setParameter(parameter);
-		model[model_index].setRef(": Not considered");
+    switch (int(input_variable[iv["iGrainBoundarySweeping"]].getValue()))
+    {
+    case 0:
+    {
+        std::vector<double> parameter;
+        parameter.push_back(0.0);
+        model[model_index].setParameter(parameter);
+        model[model_index].setRef(": Not considered");
 
-		break;
-	}
+        break;
+    }
 
-	case 1:
-	{
-		/**
-		 * @brief iGrainBoundarySweeping = 1 considers the fraction of grain swept volume (dV/V = 3 dr / r).
-		 * Then, the fraction of intra-granular gas concentration swept is dC / C = - 3 dr / r
-		 *  
-		 */
+    case 1:
+    {
+        /**
+         * @brief iGrainBoundarySweeping = 1 considers the fraction of grain swept volume (dV/V = 3 dr / r).
+         * Then, the fraction of intra-granular gas concentration swept is dC / C = - 3 dr / r
+         *  
+         */
 
-		std::vector<double> parameter;
-		/// @param[out] grain_sweeped_volume
-		parameter.push_back(3 * sciantix_variable[sv["Grain radius"]].getIncrement() / sciantix_variable[sv["Grain radius"]].getFinalValue());
-		model[model_index].setParameter(parameter);
-		model[model_index].setRef(": TRANSURANUS model");
+        std::vector<double> parameter;
+        /// @param[out] grain_sweeped_volume
+        parameter.push_back(3 * sciantix_variable[sv["Grain radius"]].getIncrement() / sciantix_variable[sv["Grain radius"]].getFinalValue());
+        model[model_index].setParameter(parameter);
+        model[model_index].setRef(": TRANSURANUS model");
 
-		break;
-	}
+        break;
+    }
 
-	default:
-		ErrorMessages::Switch(__FILE__, "iGrainBoundarySweeping", int(input_variable[iv["iGrainBoundarySweeping"]].getValue()));
-		break;
-	}
+    default:
+        ErrorMessages::Switch(__FILE__, "iGrainBoundarySweeping", int(input_variable[iv["iGrainBoundarySweeping"]].getValue()));
+        break;
+    }
+
+    MapModel();
+
+    // dC / df = - C
+
+    if (!input_variable[iv["Grain-boundary sweeping"]].getValue()) return;
+
+    // intra-granular gas diffusion modes
+    switch (int(input_variable[iv["iDiffusionSolver"]].getValue()))
+    {
+        case 1:
+        {
+            for (int i = 0; i < n_modes; ++i)
+            {
+                modes_initial_conditions[6 * 40 + i] =
+                    solver.Decay(
+                        modes_initial_conditions[6 * 40 + i],
+                        1.0,
+                        0.0,
+                        model[sm["Grain-boundary sweeping"]].getParameter().at(0)
+                    );
+            }
+            
+            break;
+        }
+
+        case 2:
+        {
+            for (int i = 0; i < n_modes; ++i)
+            {
+                modes_initial_conditions[7 * 40 + i] =
+                    solver.Decay(
+                        modes_initial_conditions[7 * 40 + i],
+                        1.0,
+                        0.0,
+                        model[sm["Grain-boundary sweeping"]].getParameter().at(0)
+                    );
+
+                modes_initial_conditions[8 * 40 + i] =
+                    solver.Decay(
+                        modes_initial_conditions[8 * 40 + i],
+                        1.0,
+                        0.0,
+                        model[sm["Grain-boundary sweeping"]].getParameter().at(0)
+                    );
+            }
+
+            break;
+        }
+
+        case 3:
+            break;
+
+        default:
+            // ErrorMessages::Switch("Simulation.h", "iDiffusionSolver", int(input_variable[iv["iDiffusionSolver"]].getValue()));
+            break;
+    }
 }
